Move CheckoutPlatformParams to :core
COSDK-626
