+++ components-core/src/main/java/com/adyen/checkout/components/core/internal/analytics/AnalyticsPlatformParams.kt
@@ -8,23 +8,36 @@
 
 package com.adyen.checkout.components.core.internal.analytics
 
-import com.adyen.checkout.components.core.internal.util.CheckoutPlatform
-import com.adyen.checkout.components.core.internal.util.CheckoutPlatformParams
+import androidx.annotation.RestrictTo
+import androidx.annotation.VisibleForTesting
+import com.adyen.checkout.components.core.BuildConfig
 
-internal object AnalyticsPlatformParams {
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+object AnalyticsPlatformParams {
 
     @Suppress("ConstPropertyName", "ktlint:standard:property-naming")
     const val channel = "android"
 
-    val platform: String
-        get() = CheckoutPlatformParams.platform.toAnalyticsPlatform().value
+    var platform = AnalyticsPlatform.ANDROID.value
+        private set
 
-    val version: String
-        get() = CheckoutPlatformParams.version
+    var version = BuildConfig.CHECKOUT_VERSION
+        private set
 
-    private fun CheckoutPlatform.toAnalyticsPlatform(): AnalyticsPlatform = when (this) {
-        CheckoutPlatform.ANDROID -> AnalyticsPlatform.ANDROID
-        CheckoutPlatform.FLUTTER -> AnalyticsPlatform.FLUTTER
-        CheckoutPlatform.REACT_NATIVE -> AnalyticsPlatform.REACT_NATIVE
+    @Suppress("unused")
+    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+    fun overrideForCrossPlatform(
+        platform: AnalyticsPlatform,
+        version: String,
+    ) {
+        AnalyticsPlatformParams.platform = platform.value
+        AnalyticsPlatformParams.version = version
+    }
+
+    @Suppress("unused")
+    @VisibleForTesting
+    internal fun resetToDefaults() {
+        platform = AnalyticsPlatform.ANDROID.value
+        version = BuildConfig.CHECKOUT_VERSION
     }
 }
