+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/AnalyticsPlatformParams.kt
@@ -8,36 +8,23 @@
 
 package com.adyen.checkout.core.analytics.internal
 
-import androidx.annotation.RestrictTo
-import androidx.annotation.VisibleForTesting
-import com.adyen.checkout.core.BuildConfig
+import com.adyen.checkout.core.common.internal.helper.CheckoutPlatform
+import com.adyen.checkout.core.common.internal.helper.CheckoutPlatformParams
 
-@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-object AnalyticsPlatformParams {
+internal object AnalyticsPlatformParams {
 
     @Suppress("ConstPropertyName", "ktlint:standard:property-naming")
     const val channel = "android"
 
-    var platform = AnalyticsPlatform.ANDROID.value
-        private set
+    val platform: String
+        get() = CheckoutPlatformParams.platform.toAnalyticsPlatform().value
 
-    var version = BuildConfig.CHECKOUT_VERSION
-        private set
+    val version: String
+        get() = CheckoutPlatformParams.version
 
-    @Suppress("unused")
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    fun overrideForCrossPlatform(
-        platform: AnalyticsPlatform,
-        version: String,
-    ) {
-        AnalyticsPlatformParams.platform = platform.value
-        AnalyticsPlatformParams.version = version
-    }
-
-    @Suppress("unused")
-    @VisibleForTesting
-    internal fun resetToDefaults() {
-        platform = AnalyticsPlatform.ANDROID.value
-        version = BuildConfig.CHECKOUT_VERSION
+    private fun CheckoutPlatform.toAnalyticsPlatform(): AnalyticsPlatform = when (this) {
+        CheckoutPlatform.ANDROID -> AnalyticsPlatform.ANDROID
+        CheckoutPlatform.FLUTTER -> AnalyticsPlatform.FLUTTER
+        CheckoutPlatform.REACT_NATIVE -> AnalyticsPlatform.REACT_NATIVE
     }
 }
