+++ core/src/test/java/com/adyen/checkout/core/common/internal/helper/CheckoutPlatformParamsTest.kt
@@ -1,6 +1,6 @@
-package com.adyen.checkout.components.core.internal.util
+package com.adyen.checkout.core.common.internal.helper
 
-import com.adyen.checkout.components.core.BuildConfig
+import com.adyen.checkout.core.BuildConfig
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
